Add files via upload
Power Chargee Utilities, an Electric Vehicle Charging Company, is increasing EV charging services due to a surge in EV adoption. This project is to conduct a Distribution Network Analysis of the EV charging stations, Substations, and distribution lines and correlate it with the Geospatial and Weather data to draw insights for the client.